chore(ci): satisfy hosted lint guardrails
