+++ crates/tsz-checker/src/error_reporter/core/mod.rs
@@ -6,6 +6,7 @@
 
 mod annotation_literal_display;
 mod annotation_text;
+mod assignment_expressions;
 mod declared_intersection_display;
 mod diagnostic_source;
 mod excess_display;
