+++ crates/tsz-checker/src/types/property_access_type/identifier_resolution.rs
@@ -653,7 +653,7 @@ impl<'a> CheckerState<'a> {
                     let summary = self.summarize_class_chain(class_idx);
                     if summary.member_kind(property_name, false, true)
                         == Some(ClassMemberKind::MethodLike)
-                        && let Some(member_info) = summary.lookup(property_name, false, true)
+                        && let Some(member_info) = summary.member_info(property_name, false, true)
                     {
                         prop_type = member_info.type_id;
                         used_class_chain_method_type = true;
@@ -965,7 +965,7 @@ impl<'a> CheckerState<'a> {
                     && let Some(class_idx) = self.nearest_enclosing_class(access.expression)
                 {
                     let summary = self.summarize_class_chain(class_idx);
-                    if let Some(member_info) = summary.lookup(property_name, true, true) {
+                    if let Some(member_info) = summary.member_info(property_name, true, true) {
                         return self.finalize_property_access_result(
                             idx,
                             effective_write_result(member_info.type_id, Some(member_info.type_id)),
@@ -973,7 +973,7 @@ impl<'a> CheckerState<'a> {
                             false,
                         );
                     }
-                    if summary.lookup(property_name, false, true).is_some() {
+                    if summary.member_info(property_name, false, true).is_some() {
                         self.error_property_not_exist_at(
                             property_name,
                             object_type_for_access,
@@ -1340,7 +1340,7 @@ impl<'a> CheckerState<'a> {
                     }
                     if let Some(member_info) = class_chain_summary
                         .as_ref()
-                        .and_then(|summary| summary.lookup(property_name, true, true))
+                        .and_then(|summary| summary.member_info(property_name, true, true))
                     {
                         return self.finalize_property_access_result(
                             idx,
