+++ scripts/arch/arch_guard_shared.py
@@ -240,9 +240,6 @@ LINE_LIMIT_CHECKS = [
             # any regression.
             "crates/tsz-checker/src/flow/control_flow/core.rs",
             "crates/tsz-checker/src/state/type_resolution/module.rs",
-            "crates/tsz-checker/src/types/class_type/constructor.rs",
-            "crates/tsz-checker/src/types/property_access_type/resolve.rs",
-            "crates/tsz-checker/src/types/type_checking/duplicate_identifiers.rs",
         },
     ),
     (
@@ -592,7 +589,7 @@ FILE_LINE_LIMIT_CHECKS = [
         / "types"
         / "property_access_type"
         / "resolve.rs",
-        3164,
+        1035,
     ),
     (
         "Checker boundary: types/type_checking/duplicate_identifiers_helpers.rs size ratchet",
@@ -656,7 +653,7 @@ FILE_LINE_LIMIT_CHECKS = [
         / "types"
         / "type_checking"
         / "duplicate_identifiers.rs",
-        2914,
+        1992,
     ),
     (
         "Checker boundary: flow/control_flow/core.rs size ratchet",
@@ -706,7 +703,7 @@ FILE_LINE_LIMIT_CHECKS = [
         / "types"
         / "class_type"
         / "constructor.rs",
-        2690,
+        1981,
     ),
     (
         "Solver boundary: diagnostics/format/compound.rs size ratchet",
@@ -1238,7 +1235,7 @@ QUERY_BOUNDARY_COMMON_REFERENCE_COUNT_CHECKS = [
         # rather than direct `query_boundaries::common` access.
         #
         # Ratcheted down after arch-smoke caught current stacked-branch slack.
-        3212,
+        3211,
     ),
 ]
